Bugfix NaN, UX improvements (#3)
More UX improvements and bugfixes

* When starting to add a recast the tool displays NaN in the Amortization Schedule
after the uiser enters a month for the new recast but before they finish entering
an amount. Don't calculate recasts unless they are valid: the month and the
dollar amount must both be positive numbers (and month must be a whole number)

* Format the monthly payment in div with id "monthlyPayment" as US Dollars.

And when the payment dropd because of a recast, append to the div a span
element with a right arror emoji and the new payment amount. Make the
title attribute of this span the year and month number.

* Format document

* Add reset button

* Delay before updating hash

* Make hash navigation changes work